src/requirements.txt
flask
python-keycloak
jwt
configparser